+++ packages/primitives/src/date-picker/index.ts
@@ -1,0 +1,62 @@
+/**
+ * Copyright IBM Corp. 2026
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+export { DatePickerStateMachine } from './machine.js';
+export { DatePickerState, DatePickerEvent } from './states.js';
+export { guards, getGuard, checkGuard } from './guards.js';
+export { actions, getAction, executeAction } from './actions.js';
+export { effects, getEffect, executeEffect } from './effects.js';
+export {
+  ClickOutsideHandler,
+  type ClickOutsideConfig,
+} from './click-outside-handler.js';
+export {
+  mapKeyboardToStateMachineEvent,
+  type KeyboardEventInfo,
+  type KeyboardEventResult,
+} from './keyboard-utils.js';
+export {
+  dateToPlainDate,
+  plainDateToDate,
+  plainDateToISOString,
+  parseISOToPlainDate,
+  parseDateToPlainDate,
+  comparePlainDates,
+  isDateInRange,
+  formatPlainDate,
+  getToday,
+  addDays,
+  addMonths,
+  daysBetween,
+  areDatesEqual,
+  getMonthStart,
+  getMonthEnd,
+  isToday,
+  isPast,
+  isFuture,
+  parseDateString,
+  isTemporalAvailable,
+  getDateHandler,
+} from './temporal-utils.js';
+export type {
+  DatePickerContext,
+  DatePickerMode,
+  InputType,
+  DateSelectPayload,
+  InputFocusPayload,
+  KeyboardPayload,
+  ValueChangePayload,
+  ValidationErrorPayload,
+  DatePickerEvent as DatePickerEventType,
+  StateTransition,
+  StateGuard,
+  StateAction,
+  SideEffect,
+  TransitionListener,
+  StateConfig,
+  TransitionMap,
+} from './types.js';