+++ packages/web-components/src/components/date-picker/index.ts
@@ -1,0 +1,15 @@
+/**
+ * Copyright IBM Corp. 2021, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+// Web Components implementation
+import './components/date-picker';
+import './components/date-picker-input';
+import './components/date-picker-input-skeleton';
+import './components/calendar-renderer';
+
+// Export state machine from primitives package
+export * from '@carbon-labs/primitives/date-picker';