+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/__stories__/HybridIPaasHeader.stories.ts
@@ -10,7 +10,7 @@
 
 import type { Meta, StoryObj } from '@storybook/web-components-vite';
 import { initialize, mswLoader } from 'msw-storybook-addon';
-import { http, HttpResponse } from 'msw';
+import { delay, http, HttpResponse } from 'msw';
 import { HybridIpaasHeader } from '../HybridIpaasHeader';
 import { html } from 'lit';
 import {
@@ -76,9 +76,13 @@ export const Basic: Story = {
   parameters: {
     msw: {
       handlers: [
-        http.get('http://localhost:6007/hybrid-ipaas/v1/header/options', () => {
-          return HttpResponse.json(mockHeaderOptions);
-        }),
+        http.get(
+          'http://localhost:6007/hybrid-ipaas/v1/header/options',
+          async () => {
+            await delay();
+            return HttpResponse.json(mockHeaderOptions);
+          }
+        ),
       ],
     },
   },
@@ -97,9 +101,13 @@ export const BasicWithTrial: Story = {
   parameters: {
     msw: {
       handlers: [
-        http.get('http://localhost:6007/hybrid-ipaas/v1/header/options', () => {
-          return HttpResponse.json({ ...mockHeaderOptions, trialConfigs });
-        }),
+        http.get(
+          'http://localhost:6007/hybrid-ipaas/v1/header/options',
+          async () => {
+            await delay();
+            return HttpResponse.json({ ...mockHeaderOptions, trialConfigs });
+          }
+        ),
       ],
     },
   },
@@ -117,9 +125,13 @@ export const Platform: Story = {
   parameters: {
     msw: {
       handlers: [
-        http.get('http://localhost:6007/hybrid-ipaas/v1/header/options', () => {
-          return HttpResponse.json(mockHeaderOptions);
-        }),
+        http.get(
+          'http://localhost:6007/hybrid-ipaas/v1/header/options',
+          async () => {
+            await delay();
+            return HttpResponse.json(mockHeaderOptions);
+          }
+        ),
       ],
     },
   },
@@ -193,10 +205,11 @@ const capabilityProfileFooterLinks = [
 
 const capabilityGlobalActions: GlobalActionConfig[] = [
   {
-    label: 'App Connect global action',
-    carbonIcon: 'AppConnectivity',
+    label: 'Cloud Services global action',
+    carbonIcon: 'CloudServices',
+    tooltip: 'This is a tooltip',
     onClick: () => {
-      console.log('App Connect global action clicked!');
+      console.log('Cloud Services global action clicked!');
     },
   },
 ];
@@ -241,6 +254,26 @@ export const LogoutCallback: Story = {
           console.log(
             'Logout callback triggered!'
           )}"></clabs-global-header-hybrid-ipaas>
+    </div>
+  `,
+};
+
+export const WithCookiePrefs: Story = {
+  parameters: {
+    msw: {
+      handlers: [
+        http.get('http://localhost:6007/hybrid-ipaas/v1/header/options', () => {
+          return HttpResponse.json(mockHeaderOptions);
+        }),
+      ],
+    },
+  },
+  render: () => html`
+    <div role="main">
+      <clabs-global-header-hybrid-ipaas
+        productName="App Connect"
+        productKey="appconnect"
+        addCookiePreferences></clabs-global-header-hybrid-ipaas>
     </div>
   `,
 };
@@ -308,9 +341,13 @@ export const RefreshOptions: Story = {
   parameters: {
     msw: {
       handlers: [
-        http.get('http://localhost:6007/hybrid-ipaas/v1/header/options', () => {
-          return HttpResponse.json(mockHeaderOptions);
-        }),
+        http.get(
+          'http://localhost:6007/hybrid-ipaas/v1/header/options',
+          async () => {
+            await delay();
+            return HttpResponse.json(mockHeaderOptions);
+          }
+        ),
       ],
     },
   },