+++ packages/web-components/src/components/global-header/components/global-header/src/globals/utils.ts
@@ -12,9 +12,13 @@ import { html } from 'lit';
 // import * as CarbonIcons from '@carbon/icons/es';
 import {
   AiLaunch20,
+  Api16,
   AppConnectivity20,
   ArrowRight16,
   Checkmark16,
+  CloudServices20,
+  ColorPalette16,
+  Cookie16,
   Document16,
   Email16,
   Help20,
@@ -60,9 +64,13 @@ export const renderCarbonIcon = (
 
   const CarbonIcons = {
     AiLaunch20,
+    Api16,
     AppConnectivity20,
     ArrowRight16,
     Checkmark16,
+    CloudServices20,
+    ColorPalette16,
+    Cookie16,
     Document16,
     Email16,
     Help20,