+++ packages/web-components/src/components/global-header/components/global-header/src/types/Header.types.ts
@@ -138,6 +138,7 @@ export type GlobalActionConfig = {
   label: string;
   carbonIcon: string;
   onClick?: () => void;
+  tooltip?: string;
 };
 
 export interface ManagementConsole {
@@ -184,11 +185,11 @@ export interface SideNav {
   buttonLabel: string;
   sidebarLabel: string;
   isCollapsible?: boolean;
-  isRail: boolean;
+  isRail?: boolean;
   onClick?: (e: Event) => void;
-  links: SideNavLink[];
+  links?: SideNavLink[];
   groups?: GroupLinks[];
-  isChildOfHeader: boolean;
+  isChildOfHeader?: boolean;
   autoCollapseOnLeave?: boolean;
   buttonCloseLabel?: string;
 }