+++ packages/web-components/src/components/global-header/vite.config.js
@@ -20,6 +20,7 @@ export default defineConfig({
     rollupOptions: {
       input: {
         'wc-global-header': 'index.ts',
+        'wc-global-header-react': 'index-react.ts',
         styles: 'components/global-header/src/index.scss',
       },
       external: ['react', 'react-dom'],